修改bot字段类型,预留空间

+++ database/migrations/2018_09_21_071634_update_article_atlas_tables.php
@@ -14,7 +14,7 @@ class UpdateArticleAtlasTables extends Migration
     public function up()
     {
         Schema::table('articles', function (Blueprint $table) {
-            $table->unsignedTinyInteger('bot')
+            $table->unsignedInteger('bot')
                 ->default(0)
                 ->comment('爬虫机器人id；0=人工编辑； 大于0=爬虫抓取');
         });
